roachpb: migrate BytesNext to encoding
* BytesNext now lives in the encoding package.
* tree now uses encoding.BytesNext instead of roachpb.Key.Next

Release note: None
