+++ pkg/sql/sem/tree/BUILD.bazel
@@ -119,7 +119,6 @@ go_library(
     deps = [
         "//pkg/geo",
         "//pkg/geo/geopb",
-        "//pkg/roachpb",
         "//pkg/security",
         "//pkg/server/telemetry",
         "//pkg/sql/catalog/catconstants",
@@ -143,6 +142,7 @@ go_library(
         "//pkg/util/bitarray",
         "//pkg/util/cache",
         "//pkg/util/duration",
+        "//pkg/util/encoding",
         "//pkg/util/errorutil/unimplemented",
         "//pkg/util/hlc",
         "//pkg/util/ipaddr",
@@ -195,7 +195,6 @@ go_test(
         "role_spec_test.go",
         "table_name_test.go",
         "time_test.go",
-        "timeconv_test.go",
         "type_check_internal_test.go",
         "type_check_test.go",
         "type_name_test.go",
@@ -206,8 +205,6 @@ go_test(
     deps = [
         "//pkg/build/bazel",
         "//pkg/internal/rsg",
-        "//pkg/kv",
-        "//pkg/roachpb",
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/settings/cluster",
@@ -234,7 +231,6 @@ go_test(
         "//pkg/util/log",
         "//pkg/util/pretty",
         "//pkg/util/randutil",
-        "//pkg/util/stop",
         "//pkg/util/timeofday",
         "//pkg/util/timetz",
         "//pkg/util/timeutil",
